% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrelr.R
\name{is_barrel}
\alias{is_barrel}
\title{Barrel Classifier}
\usage{
is_barrel(ev, la)
}
\arguments{
\item{ev}{Numeric vector. Exit velocity (in mph).}

\item{la}{Numeric vector. Launch angle (in degrees).}
}
\value{
A logical vector of the same length as the inputs.
`TRUE` indicates the (ev, la) pair is classified as a barrel.
}
\description{
Identifies whether a batted-ball event meets barrel criteria using its exit
velocity and launch angle.
}
\details{
This function determines whether the given (ev, la) pair meets the
requirements for a barreled ball according to Division I College Baseball
data. The requirements to classify a batted-ball as a barrel are:
xBA >= 0.500 & xSLG >= 1.500.
}
\examples{
is_barrel(101, 29)
is_barrel(c(103, 107), c(34, 19))

}
